change to table.js
